requirements.txt
# Core
pytz==2022.4
Django==4.1.2
django-configurations==2.4
gunicorn==20.1.0
newrelic==8.2.0.181

# For the persistence stores
psycopg2-binary==2.9.3
dj-database-url==1.0.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
Markdown==3.3.7
django-filter==22.1

# Developer Tools
ipdb==0.13.9
ipython==8.5.0
mkdocs==1.4.0
flake8==5.0.4

# Testing
mock==4.0.3
factory-boy==3.2.1
django-nose==1.4.7
nose-progressive==1.5.2
coverage==6.5.0

# Static and Media Storage
django-storages==1.13.1
boto3==1.24.85

